requirements.txt
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
decorator==4.4.2
Flask==3.1.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
instagrapi==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
moviepy==1.0.3
numpy==2.3.2
pillow==11.3.0
proglog==0.1.12
pycryptodomex==3.18.0
pydantic==1.10.9
PySocks==1.7.1
python-dotenv==1.1.1
requests==2.32.5
tqdm==4.67.1
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
yt-dlp==2025.8.22
